Fix compare URL used on pre-release versions
